backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.9.0
python-dotenv==1.0.1
pymupdf==1.24.0
pytesseract==0.3.13
sentence-transformers==3.0.0
faiss-cpu==1.8.0
numpy==1.26.4
scikit-learn==1.5.0
Pillow==10.4.0
google-generativeai>=0.8.0
